requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
distro==1.9.0
elastic-transport==8.17.1
elasticsearch==8.11.0
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.10.1
numpy==2.3.4
openai==2.6.1
packaging==25.0
pandas==2.3.3
pillow==12.0.0
portalocker==3.2.0
protobuf==6.33.0
pyarrow==21.0.0
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
qdrant-client==1.15.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.51.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typesense==1.1.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
